Backend/Services/schedule-athlete-test-service/requirements.txt
Flask==2.0.3
boto3==1.21.18
python-decouple==3.6